fix(frontend): 收敛信息中心未完成入口
拆分消息来源名称枚举，保持模型文件满足仓库行数上限。

微信设置仅暴露已接入的公众号推文渠道，并为 Debug HTTP 日志补充 URI 脱敏回归。
