+++ docs/CHANGELOG.md
@@ -25,12 +25,15 @@
 
 - 主导航页面接入统一 `FluentSurface` 视觉容器，完善首页、教务中心、课表、信息中心、学校邮箱和快速跳转页的 Fluent 2 阴影、描边、悬停与入场动效
 - 信息中心将刷新、搜索和筛选操作集中为同一操作区，快速跳转页将搜索框和最佳匹配入口合并为统一操作 surface，降低页面操作分散感
+- 设置页微信渠道仅展示已接入的微信公众号推文入口，微信服务号历史枚举保留为旧缓存兼容，不再作为可开启渠道显示
+- 拆分信息中心消息来源名称枚举，保持消息模型主文件低于仓库 500 行结构上限
 
 ### 修复
 
 - 修复 Windows WebView2 自定义用户数据目录下，微信公众平台扫码登录后 CookieManager 读取默认存储导致无法自动提取 Cookie 的问题
 - 修复第二课堂学分解析对“得分 / 分数 / 认定分”列、空白类别延续和学年学期字段识别不完整的问题
 - 学工报表 SSO、首页或第二课堂明细页返回登录页时，会在已保存 OA 密码的前提下自动刷新 OA/CAS 会话并重试
+- Debug HTTP 日志改为只输出 scheme、host 和 path，避免查询参数、fragment 或 userInfo 中的临时凭据进入控制台
 
 ### 依赖
