+++ docs/DESIGN.md
@@ -193,6 +193,7 @@ _initApp()
 - 官网消息刷新与微信推文刷新
 - 刷新进度条与单例刷新状态保持
 - 本地缓存持久化与已读状态管理
+- 设置页仅展示已接入抓取链路的渠道；微信服务号历史枚举仅用于旧缓存兼容，不作为可开启入口
 
 **UI 结构**：操作、刷新进度、搜索和筛选集中在同一个 `FluentSurface` 中，消息列表和空状态使用独立 surface，减少分散按钮和裸列表带来的层级割裂。
 
@@ -492,9 +493,9 @@ ContentDialog: 输入当前密码
 
 ---
 
-## 9. 待实现功能清单
+## 9. 后续路线
 
-后续重点包括 Android / 桌面 Release 持续集成完善、信息中心抓取源扩充、主页可配置摘要面板、快速跳转用户自定义与排序、多语言支持、数据导出 / 备份，以及安装器与签名公证完善。
+长期路线包括信息中心抓取源扩充、主页可配置摘要面板、快速跳转用户自定义与排序、多语言支持、数据导出 / 备份，以及安装器与签名公证完善。当前版本不在界面中暴露未接入渠道，未完成能力只保留在路线说明中。
 
 ---
 
@@ -507,5 +508,6 @@ ContentDialog: 输入当前密码
 5. **状态文件本地化**：桌面端保存在用户目录，移动端保存在系统分配的应用支持目录
 6. **网络请求仅用于内容抓取**：当前版本会访问学校官网与微信公众平台，不上传用户业务数据到自建服务
 7. **认证材料最小暴露**：公众号平台 Cookie / Token 仅保存在本地状态文件，不进入仓库
-8. **发布签名不入库**：Android release keystore 通过本地文件或 CI Secrets 注入
-9. **无敏感信息调试日志**：密码、教务凭据与微信认证敏感字段不输出到控制台
+8. **调试日志脱敏**：HTTP Debug 日志仅输出 scheme、host 和 path，不输出 query、fragment 或 userInfo
+9. **发布签名不入库**：Android release keystore 通过本地文件或 CI Secrets 注入
+10. **无敏感信息调试日志**：密码、教务凭据与微信认证敏感字段不输出到控制台
