+++ lib/models/channel_config_wechat.dart
@@ -3,7 +3,7 @@ part of 'channel_config.dart';
 // ==================== 微信渠道 ====================
 
 /// 微信渠道配置列表
-/// wechat_public 通过公众号平台获取已关注公众号的推文
+/// 仅暴露已经接入抓取链路的微信推文渠道，避免设置页出现不可用入口。
 const List<ChannelConfig> wechatChannels = [
   ChannelConfig(
     id: 'wechat_public',
@@ -14,12 +14,5 @@ const List<ChannelConfig> wechatChannels = [
     implemented: true,
     defaultInterval: 120,
     defaultEnabled: true,
-  ),
-  ChannelConfig(
-    id: 'wechat_service',
-    name: '微信服务号',
-    description: '暂未接入',
-    icon: FluentIcons.chat,
-    group: ChannelGroup.wechat,
   ),
 ];
